Organize directory for each table.
